move kademlia.Dialer into dial package
