+++ pkg/dial/dialer.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package kademlia
+package dial
 
 import (
 	"context"
@@ -10,6 +10,7 @@ import (
 	"go.uber.org/zap"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/peer"
+	monkit "gopkg.in/spacemonkeygo/monkit.v2"
 
 	"storj.io/storj/internal/sync2"
 	"storj.io/storj/pkg/identity"
@@ -18,20 +19,22 @@ import (
 	"storj.io/storj/pkg/transport"
 )
 
-// Dialer is a kademlia dialer
+var mon = monkit.Package()
+
+// Dialer is a dialer
 type Dialer struct {
 	log       *zap.Logger
 	transport transport.Client
 	limit     sync2.Semaphore
 }
 
-// Conn represents a kademlia connection
+// Conn represents a connection
 type Conn struct {
 	conn   *grpc.ClientConn
 	client pb.NodesClient
 }
 
-// NewDialer creates a dialer for kademlia.
+// NewDialer creates a dialer.
 func NewDialer(log *zap.Logger, transport transport.Client) *Dialer {
 	dialer := &Dialer{
 		log:       log,
